refactor: split the file into smaller
