+++ packages/react/src/components/Button/lib/variantToBackground.tsx
@@ -1,0 +1,19 @@
+import { unreachable } from '../../../_lib'
+import { Variant } from '..'
+
+export function variantToBackground(variant: Variant) {
+  switch (variant) {
+    case 'Overlay':
+      return 'surface4'
+    case 'Default':
+      return 'surface3'
+    case 'Primary':
+      return 'brand'
+    case 'Navigation':
+      return 'surface6'
+    case 'Danger':
+      return 'assertive'
+    default:
+      return unreachable(variant)
+  }
+}
